+++ tests/aave_v3/test_aave_v3_deployment.py
@@ -1,6 +1,7 @@
 """Aave v3 deployment tests."""
 import pytest
 from web3 import Web3
+from web3.contract import Contract
 
 from eth_defi.aave_v3.deployer import AaveDeployer
 from eth_defi.trace import assert_transaction_success_with_explanation
@@ -30,26 +31,17 @@ def test_aave_v3_deployment_smoke_test(
     aave_deployment: AaveDeployer,
     web3: Web3,
     user: str,
+    faucet: Contract,
+    usdc: Contract,
 ):
     """Deploy Aave against local and check something happens."""
-
-    assert web3.eth.block_number > 20
-
-    faucet = aave_deployment.get_contract_at_address(
-        web3,
-        "periphery-v3/contracts/mocks/testnet-helpers/Faucet.sol/Faucet.json",
-        "Faucet",
-    )
-    usdc = aave_deployment.get_contract_at_address(
-        web3,
-        "core-v3/contracts/mocks/tokens/MintableERC20.sol/MintableERC20.json",
-        "USDC",
-    )
+    # assert web3.eth.block_number > 20
+    assert usdc.functions.balanceOf(user).call() == 0
 
     # Get some test money
     tx_hash = faucet.functions.mint(usdc.address, user, 500 * 10**6).transact()
     assert_transaction_success_with_explanation(web3, tx_hash)
-    assert usdc.functions.balanceOf(user).call() > 0
+    assert usdc.functions.balanceOf(user).call() == 500 * 10**6
 
     # Get Aave Pool singleton
     pool = aave_deployment.get_contract_at_address(
@@ -61,14 +53,8 @@ def test_aave_v3_deployment_smoke_test(
 
 
 def test_aave_v3_deployment_smoke_test_2(
-    aave_deployment: AaveDeployer,
-    web3: Web3,
     user: str,
+    usdc: Contract,
 ):
     """Check deployer properly resets between tests."""
-    usdc = aave_deployment.get_contract_at_address(
-        web3,
-        "core-v3/contracts/mocks/tokens/MintableERC20.sol/MintableERC20.json",
-        "USDC",
-    )
     assert usdc.functions.balanceOf(user).call() == 0